requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
gunicorn==20.1.0
azure-common==1.1.3
azure-nspkg==1.0.0
azure-storage-blob==12.8.1
azure-storage==0.31.0
azure-cosmos==4.2.0
python-dotenv==1.0.0